sns-testing: add subcommand to complete the SNS swap
Currently, 'sns-testing' is only capable of deploying hardcoded SNS with
the provided SNS-controlled canister. This makes it impossible to tweak
SNS settings without recompiling 'sns-testing'.

The suggested way to create SNSes is to use 'dfx sns propose' command.
To support the workflow to deploy SNSes using this command,
'sns-testing' CLI was extended with the subcommand to complete the swap
for the created SNS.

'sns-testing run swap-complete' will wait for swap to be open, generate
required number of participation with sufficient amount of tokens to
complete the swap, and wait for swap to be completed.